frameworks/domain.core/src/main/java/yes/tdd/framewors/domain/core/GlobalIdentityService.java
package yes.tdd.framewors.domain.core;

import lombok.AccessLevel;
import lombok.NoArgsConstructor;

@NoArgsConstructor(access = AccessLevel.PRIVATE)
public class GlobalIdentityService {
    private static IdentityService identityService;

    public static String next() {
        return identityService.next();
    }

    public static String nextReadable() {
        return identityService.nextReadable();
    }

    static void reset(IdentityService identityService) {
        GlobalIdentityService.identityService = identityService;
    }
}
